+++ js/data.js
@@ -1,5 +1,4 @@
 const DESCRIPTIONS = ['Ну это я','А это не я', 'А это кто?'];
-
 const MESSAGES = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -9,7 +8,6 @@ const MESSAGES = [
   'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
   'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
 ];
-
 const NAMES = ['Борян', 'Стас', 'Линч', 'Дэвид', 'Клод', 'Валерий', 'Виталямбус'];
 
 export { DESCRIPTIONS, MESSAGES, NAMES };
